Move contact test to settings suite
