+++ app/src/androidTest/java/journal/gratitude/com/gratitudejournal/ui/SettingsFragmentInstrumentedTest.kt
@@ -1,0 +1,85 @@
+package journal.gratitude.com.gratitudejournal.ui
+
+import android.app.Activity
+import android.app.Instrumentation
+import android.content.Context
+import android.content.Intent
+import android.os.Build
+import androidx.preference.R as PreferenceR
+import androidx.recyclerview.widget.RecyclerView
+import androidx.test.core.app.ApplicationProvider
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.action.ViewActions.click
+import androidx.test.espresso.intent.Intents
+import androidx.test.espresso.intent.matcher.IntentMatchers.anyIntent
+import androidx.test.espresso.intent.matcher.IntentMatchers.hasAction
+import androidx.test.espresso.intent.matcher.IntentMatchers.hasExtra
+import androidx.test.espresso.intent.rule.IntentsRule
+import androidx.test.espresso.matcher.ViewMatchers.hasDescendant
+import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.matcher.ViewMatchers.withText
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import dagger.hilt.android.testing.HiltAndroidRule
+import dagger.hilt.android.testing.HiltAndroidTest
+import journal.gratitude.com.gratitudejournal.R
+import journal.gratitude.com.gratitudejournal.testUtils.launchFragmentInHiltContainer
+import journal.gratitude.com.gratitudejournal.ui.settings.SettingsFragment
+import org.hamcrest.CoreMatchers.allOf
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@HiltAndroidTest
+@RunWith(AndroidJUnit4::class)
+class SettingsFragmentInstrumentedTest {
+
+    @get:Rule
+    var hiltRule = HiltAndroidRule(this)
+
+    @get:Rule
+    val intentsRule = IntentsRule()
+
+    @Before
+    fun init() {
+        hiltRule.inject()
+    }
+
+    @Test
+    fun settingsFragment_clickingContactUs_opensContact() {
+        launchFragmentInHiltContainer<SettingsFragment>()
+
+        val intent = Intent()
+        val intentResult = Instrumentation.ActivityResult(Activity.RESULT_OK, intent)
+        Intents.intending(anyIntent()).respondWith(intentResult)
+
+        onView(withId(PreferenceR.id.recycler_view))
+            .perform(
+                androidx.test.espresso.contrib.RecyclerViewActions.scrollTo<RecyclerView.ViewHolder>(
+                    hasDescendant(withText(R.string.contact_us))
+                )
+            )
+        onView(withText(R.string.contact_us)).perform(click())
+
+        val context = ApplicationProvider.getApplicationContext<Context>()
+        val packageName = context.packageName
+        val packageInfo = context.packageManager.getPackageInfo(packageName, 0)
+        val emails = arrayOf("gratitude.journal.app@gmail.com")
+        val subject = context.getString(R.string.contact_us_email_subject)
+        val text = context.getString(
+            R.string.contact_us_email_body,
+            Build.MODEL,
+            Build.VERSION.RELEASE,
+            packageInfo.versionName
+        )
+
+        Intents.intended(
+            allOf(
+                hasAction(Intent.ACTION_SENDTO),
+                hasExtra(Intent.EXTRA_EMAIL, emails),
+                hasExtra(Intent.EXTRA_SUBJECT, subject),
+                hasExtra(Intent.EXTRA_TEXT, text)
+            )
+        )
+    }
+}
